pyproject.toml
[project]
name = "langgraph-deep-research-agent"
version = "0.1.0"
description = "An autonomous daily research agent that discovers, filters, and delivers research updates using LangGraph subagent delegation."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "tavily-python>=0.7.23",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/deep_research_agent"]

[tool.interrogate]
fail-under=100
exclude = ["tests", "build", "dist", "tmp", ".venv", ".ruff_cache"]
ignore-module = true
ignore-init-method = true
ignore-regex = ["^_execute_callbacks"]
verbose = 2
omit-covered-files = true

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = ["__pycache__"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "Q",    # flake8-quotes
    "A",    # prevent using keywords that clobber python builtins
    "ISC",  # implicit string concatenation
    "UP",   # alert you when better syntax is available in your python version, pyupgrade
    "RUF",  # the ruff developer's own rules
    "D",    # pydocstyle
    "I",    # isort
    "ANN", # flake8-annotations - Missing type annotation for function argument
    "T20", # flake8-print - print and pprint are not allowed
]
ignore = [
    "E501",
    "A001",  # Ignore shadowing of python builtin words
    "D100",  # Missing docstring in public module
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "D104",  # Missing docstring in public package
    "D105",  # Missing docstring in magic method
    "D205",  # 1 blank line required between summary line and description
    "D209",  # Multi-line docstring
    "D210",  # No whitespaces allowed surrounding docstring text
    "D400",  # First line should end with a period
    "D401",  # First line should be in imperative mood
    "D404",  # First word of the docstring should not be `This`
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["ANN"]
"examples/*.py" = ["ANN", "T20"]
"examples/*.ipynb" = ["ANN", "T20", "A004"]

# Docstring formatting options
[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 100

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
exclude = ["tests"]

reportMissingTypeArgument = "warning"
reportUnnecessaryCast = "warning"
reportUnusedIgnoreComment = "warning"
reportUnreachable = "warning"
reportUnknownMemberType = "warning"
reportUnknownArgumentType = "warning"

[[tool.pyright.executionEnvironments]]
root = "tests"
typeCheckingMode = "basic"
reportMissingTypeArgument = "warning"
reportUnknownParameterType = "warning"
reportMissingParameterType = "warning"
reportUnknownArgumentType = "warning"
reportPrivateUsage = "warning"
